+++ faster_coco_eval/utils/pytorch/__init__.py
@@ -1,0 +1,4 @@
+from .coco_dataset import FasterCocoDetection
+from .coco_eval import FasterCocoEvaluator
+
+__all__ = ["FasterCocoEvaluator", "FasterCocoDetection"]